Adding support to run browser windowless in tests
It makes tests run inside `nix-shell` on non-NixOS installations.
Unfortunately accessing OpenGL libs from Nix packages are an issue on
other distributions.

It could be a nice extension to the application itself to have a CLI
switch to activate windowless mode. However as the browser accesses `sys.argv`
(questionable design choice) it is theoretically possible to run it
windowless.